requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt==4.0.1
python-multipart
passlib[bcrypt]
pyjwt
python-jose
docker

